+++ Grading-System/src/services/CourseService.java
@@ -15,7 +15,7 @@ public class CourseService {
     }
 
     // Create a course and add it to the system
-    public void createCourse(String name, String courseId, int creditHours) {
+    public void createCourse(String name, String courseId, double creditHours) {
         courses.put(courseId, new Course(name, courseId, creditHours));
     }
